Reorganize runtime outcome script under scripts with compatibility shim
